refactor to pull data fetching to top
